requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
openai==1.10.0
resend==0.7.0
beautifulsoup4==4.12.3
aiofiles==23.2.1
